requirements.txt
fastapi>=0.95.0,<0.96.0
uvicorn>=0.21.1,<0.22.0
joblib>=1.1.0,<1.2.0
scikit-learn>=1.2.2,<1.3.0
numpy>=1.24.3,<1.25.0
pandas>=2.0.0,<2.1.0
requests>=2.30.0,<2.31.0
python-dotenv>=1.0.0,<1.1.0